explore checks and repair of minimum mesh size violation
near-coincident grid lines cause diverging execution time and invalid results.
Use genParamMinGridSpacingX, genParamMinGridSpacingY and genParamMinGridSpacingZ values previously introduced to check for violations and remove lines that are too close together. (Older versions of Octave that still run on Win7 do not have built-in uniquetol() - in that case it needs to be added to the script directory. Get it here: https://searchcode.com/codesearch/view/74351304/9.

This probably isn't the best implementation, as it relies on the use of DefineRectGrid(). but highlights that mesh quality checks implemented in the .m script can have an immediate impact on user friendliness, as they give tangible feedback. The cause of redundant or closely-spaced grid lines is usually the use of nested grid definitions.